ai-service/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
numpy
opencv-python
httpx
jinja2
